skriver ting til tsx
